requirements.txt
requests==2.31.0
python-dotenv==1.0.0
selenium==4.16.0
beautifulsoup4==4.12.2
pillow==10.0.0
fake-useragent==1.4.0
loguru==0.7.2
docker==7.0.0
pytest==7.4.3
pytest-cov==4.1.0 